Remove clap dependency from turbopack-cli-utils
